+++ test/zprint/core_test.clj
@@ -4,7 +4,7 @@
    [expectations :refer :all]
    [zprint.core :refer :all]
    [zprint.zprint :refer :all]
-   [zprint.config :refer :all]
+   [zprint.config :refer :all :exclude [set-options! configure-all! get-options]]
    [clojure.repl :refer :all]))
 
 ;;