+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # Change Log
 All notable changes to this project will be documented in this file. 
-## [Unreleased]
+
+## 0.2.1 - 2016-09-22
 ### Changed
+
+* Moved some configuration functions into zprint.core to match readme.
+* Put ALPHA banner in readme.
 
 ### Removed
 
 ### Fixed
 
+* Fixed a bug where it wouldn't configure if no ~/.zprintrc file.
+
 ## 0.2.0 - 2016-09-21
 ### Added
 - Initial project commit.
 
-[Unreleased]: https://github.com/kkinnear/zprint/compare/0.1.1...HEAD